Added crude emulator for testing
Defined an emulator for the OC110 when accessed over the serial port.
The emulation is far from complete at the moment, but we've got a full
dump of a session against the OC110, so it shouldn't be too hard to add
emulation of enough commands to make testing and further development
pretty simple without having to have continual physical access to a
unit.