Lots of new stuff!
Added Sample and SampleHead classes to handle parsing and reconstructing
the data-feed from the OC110 which represents sample bottles and their
measuring heads.

Added a DataLogger class which provides pre-packaged methods
representing the commands that can be sent to the unit, and handles
parsing the results (including ensuring checksums match where provided).

Emulator still needs more work to cover the complete set of data we
retrieved from the logging session.