Lots of work on exporting
Finished export handlers for the bottle list and bottle readings
sub-windows, for the CSV and Excel formats. Added a couple of dialogs
(export_csv_dialog and export_excel_dialog) to handle prompting the user
for various export settings, and added a generic BaseExporter class (in
the new exporter module) to consolidate the common code between the two
exports.

Also added the ability to switch between absolute and delta readings in
the bottle readings sub-window, and added the missing dilution edit.