Started work on new command line apps
The new command line apps should share as much code as possible with the
GUI and each other. The first step, now done, is to separate all the
common stuff into a base command line application class (Application)
which defines a standard suite-wide option for specifying the
data-logger COM port (-p) and constructs it in the app intializer.

The list application is more or less done now too, incorporating support
for listing individual bottles or the complete set.

The dump application is yet to be written and will require refactoring
the export routines from the GUI into their own module shared by this
application.