Centralized moving average and delta functions
Moved the calculation of delta values and moving averages out of the
bottle_window GUI module and into the base bottles module in a new
DataAnalyzer class so that all applications in the suite can use the
same code. Updated the BottleModel class to use the new class as well so
the list-view matches the graphed data. Tweaked the calculation of
moving average timestamps so that we're returning timestamps in the
middle of the calculated value (scientific moving average) as opposed to
the end (financial moving average).