Added momentary readings
Split BottleReadings into BottleAutoReadings and BottleManualReadings,
the latter of which is used to retrieve so called "momentary" values
when the unit is in pressure mode. Dumped the fixed example data for the
DummyLogger class and added XML serialization of the Bottle class along
with an example XML file taken from the OC110 unit we've been
experimenting with (adjust MANIFEST.in to include it?).

Refined the string parsing and construction; might have figured out the
auto-temp adaptation phase field, and possibly the auto-temp or dilution
field, along with the construction of the last field in BOD standard
mode. Still needs bug fixing though as we've broken readings for
oxitoplist at least...