Removed manual property from DataAnalyzer
Makes no sense to interpolate manual ("momentary") values in
DataAnalyzer. As soon as manual values are introduced there's no longer
a regular time gap between values so the moving average becomes invalid
(without weighting to compensate for the different time gaps).

Better to use the manual values as simple markers on the graph and list
them separately in the table as the original software does.